Start splitting Financial function tests out from the large FinancialTests class into individual test classes for each function
