+++ tests/PhpSpreadsheetTests/Calculation/FinancialTest.php
@@ -14,134 +14,6 @@ class FinancialTest extends TestCase
     }
 
     /**
-     * @dataProvider providerAMORDEGRC
-     *
-     * @param mixed $expectedResult
-     */
-    public function testAMORDEGRC($expectedResult, ...$args): void
-    {
-        $result = Financial::AMORDEGRC(...$args);
-        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
-    }
-
-    public function providerAMORDEGRC()
-    {
-        return require 'tests/data/Calculation/Financial/AMORDEGRC.php';
-    }
-
-    /**
-     * @dataProvider providerAMORLINC
-     *
-     * @param mixed $expectedResult
-     */
-    public function testAMORLINC($expectedResult, ...$args): void
-    {
-        $result = Financial::AMORLINC(...$args);
-        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
-    }
-
-    public function providerAMORLINC()
-    {
-        return require 'tests/data/Calculation/Financial/AMORLINC.php';
-    }
-
-    /**
-     * @dataProvider providerCOUPDAYBS
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCOUPDAYBS($expectedResult, ...$args): void
-    {
-        $result = Financial::COUPDAYBS(...$args);
-        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
-    }
-
-    public function providerCOUPDAYBS()
-    {
-        return require 'tests/data/Calculation/Financial/COUPDAYBS.php';
-    }
-
-    /**
-     * @dataProvider providerCOUPDAYS
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCOUPDAYS($expectedResult, ...$args): void
-    {
-        $result = Financial::COUPDAYS(...$args);
-        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
-    }
-
-    public function providerCOUPDAYS()
-    {
-        return require 'tests/data/Calculation/Financial/COUPDAYS.php';
-    }
-
-    /**
-     * @dataProvider providerCOUPDAYSNC
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCOUPDAYSNC($expectedResult, ...$args): void
-    {
-        $result = Financial::COUPDAYSNC(...$args);
-        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
-    }
-
-    public function providerCOUPDAYSNC()
-    {
-        return require 'tests/data/Calculation/Financial/COUPDAYSNC.php';
-    }
-
-    /**
-     * @dataProvider providerCOUPNCD
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCOUPNCD($expectedResult, ...$args): void
-    {
-        $result = Financial::COUPNCD(...$args);
-        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
-    }
-
-    public function providerCOUPNCD()
-    {
-        return require 'tests/data/Calculation/Financial/COUPNCD.php';
-    }
-
-    /**
-     * @dataProvider providerCOUPNUM
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCOUPNUM($expectedResult, ...$args): void
-    {
-        $result = Financial::COUPNUM(...$args);
-        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
-    }
-
-    public function providerCOUPNUM()
-    {
-        return require 'tests/data/Calculation/Financial/COUPNUM.php';
-    }
-
-    /**
-     * @dataProvider providerCOUPPCD
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCOUPPCD($expectedResult, ...$args): void
-    {
-        $result = Financial::COUPPCD(...$args);
-        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
-    }
-
-    public function providerCOUPPCD()
-    {
-        return require 'tests/data/Calculation/Financial/COUPPCD.php';
-    }
-
-    /**
      * @dataProvider providerCUMIPMT
      *
      * @param mixed $expectedResult
